Move grant summary under Conference
