+++ backend/conferences/admin/__init__.py
@@ -1,0 +1,15 @@
+from .conference import (
+    ConferenceAdmin,
+    DeadlineAdmin,
+    AudienceLevelAdmin,
+    SpeakerVoucherAdmin,
+    TopicAdmin,
+)
+
+__all__ = [
+    "AudienceLevelAdmin",
+    "ConferenceAdmin",
+    "DeadlineAdmin",
+    "SpeakerVoucherAdmin",
+    "TopicAdmin",
+]
