+++ backend/conferences/templates/admin/conferences/conference/change_form.html
@@ -1,0 +1,9 @@
+{% extends "admin/change_form.html" %}
+{% block object-tools-items %}
+    <li>
+        <a href="{% url 'admin:grants_summary' object_id=original.id %}" class="button">
+          Grants Summary
+        </a>
+    </li>
+    {{ block.super }}
+{% endblock %}
