+++ backend/grants/templates/admin/grants/grant/change_list.html
@@ -1,11 +1,4 @@
 {% extends "admin/change_list.html" %}
-{% block object-tools-items %}
-    <li>
-        <a href="{% url 'admin:grants-summary' %}?{{ request.GET.urlencode }}"
-           class="button">Grant Summary</a>
-    </li>
-    {{ block.super }}
-{% endblock %}
 
 {% block extrahead %}
     {{ block.super }}
